+++ src/guiguts/application.py
@@ -17,7 +17,6 @@ if __name__ == "__main__" and __package__ is None:
 from guiguts.file import File, NUM_RECENT_FILES
 from guiguts.maintext import maintext
 from guiguts.mainwindow import (
-    root,
     MainWindow,
     Menu,
     menubar,
@@ -25,9 +24,9 @@ from guiguts.mainwindow import (
     statusbar,
     ErrorHandler,
 )
-
 from guiguts.page_details import PageDetailsDialog
 from guiguts.preferences import preferences
+from guiguts.root import root
 from guiguts.search import show_search_dialog, find_next
 from guiguts.spell import spell_check
 from guiguts.utilities import is_mac
@@ -393,7 +392,10 @@ Fifth Floor, Boston, MA 02110-1301 USA."""
     def init_tools_menu(self) -> None:
         """Create the Tools menu."""
         menu_edit = Menu(menubar(), "~Tools")
-        menu_edit.add_button("~Spelling Check", spell_check)
+        menu_edit.add_button(
+            "~Spelling Check",
+            lambda: spell_check(self.file.add_good_word_to_project_dictionary),
+        )
 
     def init_view_menu(self) -> None:
         """Create the View menu."""
